separate UI from main class WIP
